+++ NEWS.md
@@ -1,6 +1,7 @@
 # crunch (development version)
 * https verification can be disabled during testing by setting environment variable `R_TEST_VERIFY_SSL=FALSE`
 * You can now use a named `filter` or `filter` object when using `tabbook()`. Filtering by expression in the dataset argument is also supported again.
+* `newMultiTable()` now correctly passes `...`, so arguments like `is_public` work (#424)
 
 # crunch 1.26.3
 * Can now `deriveArray()` using expressions to create the subvariables.